requirements.txt
pygame~=2.5.0
neat-python
